requirements.txt
google-adk
fastapi
mcp
uvicorn
pydantic
python-dotenv
streamlit
